+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ## <next>
 
 * `Cards` fix collapse/expand icon container placement
+* `Menu` component added
+* `KEY_CODES` constant added
+* `PerfProfiler` utility component added
 
 ## 2.0.1